requirements.txt
Pillow
python-telegram-bot
emoji
